+++ start_emulator.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env sh
+exec firebase -P ioxio-local-dev emulators:start --only firestore